[v3.7.3 May 14th, 2025] Refactor styles and improve button loading).
Smoother Buttons: I've improved how the fast-forward and rewind buttons
load on YouTube. They should now appear more smoothly when you start a
video or click to a new one, with less of an abrupt "pop-in." This
makes them feel more integrated with the player.

Code Organization: I've reorganized some of the extension's code
"under the hood." I moved the styling rules for the popup settings
window out of the popup.html file and into the main style.css file.
This doesn't change how the popup looks to you, but it makes the code
neater and easier for me to manage for future updates.